move dark class logic to context
