+++ src/contexts/darkMode.tsx
@@ -1,4 +1,4 @@
-import { createContext, useContext, useState } from "react";
+import { createContext, useContext, useEffect, useState } from "react";
 import { type ReactNode } from "react";
 
 const DarkModeContext = createContext<{
@@ -14,6 +14,16 @@ export const DarkModeProvider = ({ children }: { children: ReactNode }) => {
   ).matches;
   const [darkModeEnabled, setDarkModeEnabled] = useState(isSystemDark);
   const toggleDarkMode = () => setDarkModeEnabled(!darkModeEnabled);
+
+  const htmlClasses = document.querySelector("html")!.classList;
+
+  useEffect(() => {
+    if (darkModeEnabled) {
+      htmlClasses.add("dark");
+    } else {
+      htmlClasses.remove("dark");
+    }
+  }, [darkModeEnabled]);
 
   return (
     <DarkModeContext.Provider value={{ darkModeEnabled, toggleDarkMode }}>
